# src/pwa/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WebAZ Protocol API",
    "version": "0.4.14",
    "description": "Auto-generated endpoint inventory (640 endpoints, 17 with full schema). See docs/ for design context."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "api_key"
      },
      "webauthnToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-WebAuthn-Token"
      }
    }
  },
  "paths": {
...
    "/api/admin/admins": {
      "get": {
        "summary": "GET 全部 admin 列表",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "POST 创建 admin（仅 root）",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/admins/:id": {
      "delete": {
        "summary": "DELETE 撤销 admin（root only；不能撤自己；至少保留 1 个 root）",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/admins/:id/permissions": {
      "patch": {
        "summary": "PATCH 更新权限（root only）",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/announcements": {
      "post": {
        "summary": "POST /api/admin/announcements",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/announcements/:id": {
      "patch": {
        "summary": "PATCH /api/admin/announcements/:id",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/announcements/:id/read": {
      "post": {
        "summary": "POST /api/announcements/:id/read",
        "tags": [],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/announcements/active": {
      "get": {
        "summary": "列出对当前用户可见的活跃公告（按角色 + 区域过滤）",
        "tags": [],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/me/delete-cancel": {
      "post": {
        "summary": "POST /api/me/delete-cancel",
        "tags": [],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/me/delete-request": {
      "post": {
        "summary": "POST /api/me/delete-request",
        "tags": [],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/me/delete-status": {
      "get": {
        "summary": "GET /api/me/delete-status",
        "tags": [],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}